correct accept rate output

+++ include/vsmc/core/sampler.hpp
@@ -462,7 +462,8 @@ class Sampler
             os << resampled_history_[iter] << sep;
             if (print_accept) {
                 for (unsigned c = 0; c != accept_history_[iter].size(); ++c)
-                    os << accept_history_[iter][c] / size() << sep;
+                    os << accept_history_[iter][c] /
+                        static_cast<double>(size()) << sep;
                 unsigned diff = static_cast<unsigned>(
                         accd - accept_history_[iter].size());
                 for (unsigned c = 0; c != diff; ++c)
